portal-service-test Move role related test classes of the com.liferay.portal.service package to roles-test module
